autoformat

+++ pycromanager/core.py
@@ -71,12 +71,12 @@ class JavaSocket:
         if self._debug:
             print("DEBUG, sending: {}".format(message))
         if timeout == 0:
-            self._socket.send(bytes(message_string, 'iso-8859-1'))
+            self._socket.send(bytes(message_string, "iso-8859-1"))
         else:
             start = time.time()
             while 1000 * (time.time() - start) < timeout:
                 try:
-                    self._socket.send(bytes(message_string, 'iso-8859-1'), flags=zmq.NOBLOCK)
+                    self._socket.send(bytes(message_string, "iso-8859-1"), flags=zmq.NOBLOCK)
                     return True
                 except zmq.ZMQError:
                     pass  # ignore, keep trying
@@ -97,7 +97,7 @@ class JavaSocket:
                     pass  # ignore, keep trying
             if reply is None:
                 return reply
-        message = json.loads(reply.decode('iso-8859-1'))
+        message = json.loads(reply.decode("iso-8859-1"))
         if self._debug:
             print("DEBUG, recieved: {}".format(message))
         self._check_exception(message)
@@ -469,7 +469,8 @@ class JavaObjectShadow:
 
 
 def serialize_array(array):
-    return array.tobytes().decode('iso-8859-1')
+    return array.tobytes().decode("iso-8859-1")
+
 
 def deserialize_array(json_return):
     """
@@ -478,19 +479,18 @@ def deserialize_array(json_return):
     ----------
     json_return
     """
-    if json_return['type'] in ['byte-array', 'int-array', 'short-array', 'float-array']:
-        decoded = bytes(json_return['value'], 'iso-8859-1')
-        if json_return['type'] == 'byte-array':
-            return np.frombuffer(decoded, dtype='>u1').copy()
-        elif json_return['type'] == 'double-array':
-            return np.frombuffer(decoded, dtype='>f8').copy()
-        elif json_return['type'] == 'int-array':
-            return np.frombuffer(decoded, dtype='>u4').copy()
-        elif json_return['type'] == 'short-array':
-            return np.frombuffer(decoded, dtype='>u2').copy()
-        elif json_return['type'] == 'float-array':
-            return np.frombuffer(decoded, dtype='>f4').copy()
-
+    if json_return["type"] in ["byte-array", "int-array", "short-array", "float-array"]:
+        decoded = bytes(json_return["value"], "iso-8859-1")
+        if json_return["type"] == "byte-array":
+            return np.frombuffer(decoded, dtype=">u1").copy()
+        elif json_return["type"] == "double-array":
+            return np.frombuffer(decoded, dtype=">f8").copy()
+        elif json_return["type"] == "int-array":
+            return np.frombuffer(decoded, dtype=">u4").copy()
+        elif json_return["type"] == "short-array":
+            return np.frombuffer(decoded, dtype=">u2").copy()
+        elif json_return["type"] == "float-array":
+            return np.frombuffer(decoded, dtype=">f4").copy()
 
 
 def _package_arguments(valid_method_spec, fn_args):
